refactor(project): moved services to correct gradle project
